Move action 2 ctrl tensor to subproc vec env
