+++ envs/legged_gym/legged_gym_env.py
@@ -5,7 +5,6 @@ from orca_gym.utils import rotations
 from typing import Optional, Any, SupportsFloat
 from gymnasium import spaces
 import datetime
-import torch
 
 from .legged_robot import LeggedRobot
 
@@ -41,14 +40,9 @@ class LeggedGymEnv(OrcaGymMultiAgentEnv):
         )
 
 
-    def step_agents(self, action: np.ndarray) -> None:
+    def step_agents(self, action: np.ndarray, actuator_ctrl: np.ndarray) -> None:
         # print("Step agents: ", action)
 
-        # 将 action 线性变换到 actuator 的控制范围空间
-        if self._use_tensor:
-            actuator_ctrl = self._action2ctrl_tensor(action)
-        else:
-            actuator_ctrl = self._action2ctrl(action)
 
         # 性能优化：在Env中批量更新所有agent的控制量
         if self._task != "no_action":
@@ -158,26 +152,7 @@ class LeggedGymEnv(OrcaGymMultiAgentEnv):
         self.set_mocap_pos_and_quat(mocaps)
         self.mj_forward()
 
-    def reorder_agents(self):
-        ctrl_info = {}
-        for agent in self._agents:
-            ctrl_info[agent.name] = agent.get_ctrl_info()
 
-        # 按照 ctrl_start 排序
-        ctrl_info = {k: v for k, v in sorted(ctrl_info.items(), key=lambda item: item[1]["ctrl_start"])}
-
-        reordered_agents = []
-        for agent_name in ctrl_info.keys():
-            for agent in self._agents:
-                if agent.name == agent_name:
-                    reordered_agents.append(agent)
-                    break
-
-        assert len(reordered_agents) == len(self._agents)
-
-        self._generate_action_scale_array(ctrl_info)
-
-        return reordered_agents
 
     def _generate_contact_dict(self) -> dict[str, list[str]]:
         contacts = self.query_contact_simple()
@@ -195,99 +170,3 @@ class LeggedGymEnv(OrcaGymMultiAgentEnv):
 
         return contact_dict
 
-    def _generate_action_scale_array(self, ctrl_info: dict) -> np.ndarray:
-        self._actuator_type = next(iter(ctrl_info.values()))["actuator_type"]           # shape = (1)
-        self._action_scale = next(iter(ctrl_info.values()))["action_scale"]             # shape = (1)
-        self._action_space_range = next(iter(ctrl_info.values()))["action_space_range"] # shape = (2)
-
-        self._ctrl_start = np.array([ctrl["ctrl_start"] for key, ctrl in ctrl_info.items()]) # shape = (agent_num)
-        self._ctrl_end = np.array([ctrl["ctrl_end"] for key, ctrl in ctrl_info.items()])     # shape = (agent_num)
-        self._ctrl_range = np.array([ctrl["ctrl_range"] for key, ctrl in ctrl_info.items()]).reshape(-1, 2)   # shape = (agent_num x actor_num, 2) 
-        self._ctrl_delta_range = np.array([ctrl["ctrl_delta_range"] for key, ctrl in ctrl_info.items()]).reshape(-1, 2)  # shape = (agent_num x actor_num, 2)
-        self._neutral_joint_values = np.array([ctrl["neutral_joint_values"] for key, ctrl in ctrl_info.items()]).reshape(-1) # shape = (agent_num x actor_num)
-
-        if len(self._agents) > 4:
-            self._use_tensor = True
-        else:
-            self._use_tensor = False
-
-        if not self._use_tensor:
-            return
-
-        # 初始化阶段，将不变数据移动到 GPU
-        # device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        device = torch.device("cpu")
-        self._action_tensor = torch.zeros(len(self._ctrl_range), device=device)  # 每个 action 的 tensor
-        self._action_scale_tensor = torch.ones(len(self._ctrl_range), device=device) * self._action_scale  # 每个 action 的缩放因子
-        self._scaled_action_tensor = torch.zeros(len(self._ctrl_range), device=device)  # 缩放后的 action
-        self._action_space_range_tensor = torch.tensor(self._action_space_range, device=device)
-        self._clipped_action_tensor = torch.zeros(len(self._ctrl_range), device=device)  # 限制在有效范围内的 action
-
-        self._ctrl_range_tensor = torch.tensor(self._ctrl_range, device=device)
-        self._ctrl_delta_range_tensor = torch.tensor(self._ctrl_delta_range, device=device)
-        self._neutral_joint_values_tensor = torch.tensor(self._neutral_joint_values, device=device)
-        self._ctrl_delta_tensor = torch.zeros(len(self._ctrl_range), device=device)  # 插值后的控制量
-        self._actuator_ctrl_tensor = torch.zeros(len(self._ctrl_range), device=device)  # actuator 控制量
-
-
-    def _action2ctrl_tensor(self, action: np.ndarray) -> np.ndarray:
-        # 将 action 传输到 GPU
-        self._action_tensor[:] = torch.from_numpy(action).to(self._action_tensor.device)
-
-        # 缩放后的 action
-        self._scaled_action_tensor = self._action_tensor * self._action_scale_tensor
-
-        # 限制 scaled_action 在有效范围内
-        self._clipped_action_tensor = torch.clamp(self._scaled_action_tensor, self._action_space_range_tensor[0], self._action_space_range_tensor[1])
-
-        # 批量计算插值
-        if (self._actuator_type == "position"):
-            self._ctrl_delta_tensor = (
-                self._ctrl_delta_range_tensor[:, 0] +  # fp1
-                (self._ctrl_delta_range_tensor[:, 1] - self._ctrl_delta_range_tensor[:, 0]) *  # (fp2 - fp1)
-                (self._clipped_action_tensor - self._action_space_range_tensor[0]) /  # (x - xp1)
-                (self._action_space_range_tensor[1] - self._action_space_range_tensor[0])  # (xp2 - xp1)
-            )
-
-            self._actuator_ctrl_tensor = self._ctrl_delta_tensor + self._neutral_joint_values_tensor
-        elif (self._actuator_type == "torque"):
-            self._actuator_ctrl_tensor = (
-                self._ctrl_range_tensor[:, 0] +  # fp1
-                (self._ctrl_range_tensor[:, 1] - self._ctrl_range_tensor[:, 0]) *  # (fp2 - fp1)
-                (self._clipped_action_tensor - self._action_space_range_tensor[0]) /  # (x - xp1)
-                (self._action_space_range_tensor[1] - self._action_space_range_tensor[0])  # (xp2 - xp1)
-            )
-        else:
-            raise ValueError(f"Unsupported actuator type: {self._actuator_type}")
-    
-        return self._actuator_ctrl_tensor.cpu().numpy()
-        
-
-    def _action2ctrl(self, action: np.ndarray) -> np.ndarray:
-        # 缩放后的 action
-        scaled_action = action * self._action_scale
-
-        # 限制 scaled_action 在有效范围内
-        clipped_action = np.clip(scaled_action, self._action_space_range[0], self._action_space_range[1])
-
-        # 批量计算插值
-        if (self._actuator_type == "position"):
-            ctrl_delta = (
-                self._ctrl_delta_range[:, 0] +  # fp1
-                (self._ctrl_delta_range[:, 1] - self._ctrl_delta_range[:, 0]) *  # (fp2 - fp1)
-                (clipped_action - self._action_space_range[0]) /  # (x - xp1)
-                (self._action_space_range[1] - self._action_space_range[0])  # (xp2 - xp1)
-            )
-
-            actuator_ctrl = self._neutral_joint_values + ctrl_delta
-        elif (self._actuator_type == "torque"):
-            actuator_ctrl = (
-                self._ctrl_range[:, 0] +  # fp1
-                (self._ctrl_range[:, 1] - self._ctrl_range[:, 0]) *  # (fp2 - fp1)
-                (clipped_action - self._action_space_range[0]) /  # (x - xp1)
-                (self._action_space_range[1] - self._action_space_range[0])  # (xp2 - xp1)
-            )
-        else:
-            raise ValueError(f"Unsupported actuator type: {self._actuator_type}")
-        
-        return actuator_ctrl
